requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymongo
jinja2
aiofiles
mammoth
beautifulsoup4
python-dotenv